feat: :art: rewrite file acceptance logic for accept object and sync tasks
