+++ storage-node/src/services/sync/acceptPendingObjects.ts
@@ -3,14 +3,12 @@ import { KeyringPair } from '@polkadot/keyring/types'
 import fs from 'fs'
 import _ from 'lodash'
 import path from 'path'
-import { addDataObjectIdToCache } from '../caching/localDataObjects'
 import { registerNewDataObjectId } from '../caching/newUploads'
 import { hashFile } from '../helpers/hashing'
-import { moveFile } from '../helpers/moveFile'
+import { acceptObject } from '../helpers/moveFile'
 import logger from '../logger'
 import { QueryNodeApi } from '../queryNode/api'
 import { acceptPendingDataObjectsBatch } from '../runtime/extrinsics'
-import { getStorageProviderConnection, isStorageProviderConnectionEnabled } from '../../commands/server'
 
 const fsPromises = fs.promises
 
@@ -75,17 +73,12 @@ export class AcceptPendingObjectsService {
       .catch(() => false)
   }
 
-  /**
-   * Retrieves the list of pending objects from the specified folder.
-   * @returns A promise that resolves to an array of strings representing the names of the pending objects.
-   */
   private async getPendingObjectsFromFolder(): Promise<string[]> {
     const dirEntries = await fsPromises.readdir(this.pendingDataObjectsDir, { withFileTypes: true })
     return dirEntries.filter((entry) => entry.isFile()).map((entry) => entry.name)
   }
 
-  // making it public just for testing
-  public runWithInterval(api: ApiPromise, workerId: number, maxTxBatchSize: number, intervalMs: number) {
+  private runWithInterval(api: ApiPromise, workerId: number, maxTxBatchSize: number, intervalMs: number) {
     const run = () => {
       this.getPendingObjectsFromFolder()
         .then((ids) => this.processPendingObjects(ids))
@@ -96,13 +89,6 @@ export class AcceptPendingObjectsService {
     run()
   }
 
-  /**
-   * Processes the pending objects with the given IDs:
-   * - deletes pending objects that have been already deleted
-   * - moves pending objects to the uploads directory only if their ipfs hash is verified and they are assigned to one of the upload buckets
-   * @param pendingIds - An array of string IDs representing the pending objects.
-   * @returns A Promise that resolves to an array of PendingObjectDetails, containing objects that need to be accepted by a storage bucket
-   */
   private async processPendingObjects(pendingIds: string[]): Promise<PendingObjectDetails> {
     const pendingDataObjects = await this.qnApi.getDataObjectDetails(pendingIds)
 
@@ -147,7 +133,7 @@ export class AcceptPendingObjectsService {
         )
         if (storageBucket) {
           if (dataObject.isAccepted) {
-            await this.moveToAcceptedLocation(dataObject.id)
+            await this.movePendingDataObjectToUploadsDir(dataObject.id)
           } else {
             objectsToAccept.push([dataObject.id, [storageBucket.storageBucket.id, dataObject.storageBag.id]])
           }
@@ -155,23 +141,14 @@ export class AcceptPendingObjectsService {
           logger.debug(
             `Data object ${dataObject.id} in pending directory is no longer assigned to any of the upload buckets: ${this.uploadBuckets}. Moving it to the uploads directory.`
           )
-          await this.moveToAcceptedLocation(dataObject.id)
+          await this.movePendingDataObjectToUploadsDir(dataObject.id)
         }
       })
     )
     return objectsToAccept
   }
 
-  /**
-   * Moves a data object to the accepted location.
-   * If the file already exists in the uploads directory, it will be deleted before moving the new file.
-   * If the storage provider connection is enabled, the file will be uploaded to the remote bucket before moving.
-   * After the file is moved, the data object ID will be registered and added to the cache.
-   * Setting it to public just for testing
-   * @param dataObjectId - The ID of the data object to move.
-   * @returns A Promise that resolves when the operation is complete.
-   */
-  public async moveToAcceptedLocation(dataObjectId: string): Promise<void> {
+  private async movePendingDataObjectToUploadsDir(dataObjectId: string): Promise<void> {
     const currentPath = path.join(this.pendingDataObjectsDir, dataObjectId)
     const newPath = path.join(this.uploadsDir, dataObjectId)
 
@@ -183,19 +160,10 @@ export class AcceptPendingObjectsService {
         await fsPromises.unlink(currentPath)
       } catch {
         // If the file does not exist in the uploads directory, proceed with the rename
-        const isFileDestinedToLocalVolume = isStorageProviderConnectionEnabled()
-        if (!isFileDestinedToLocalVolume) {
-          await moveFile(currentPath, newPath)
-        } else {
-          const connection = getStorageProviderConnection()!
-          await connection.uploadFileToRemoteBucket(dataObjectId, currentPath) // NOTE: consider converting to non blocking promise
-          await fsPromises.unlink(currentPath) // delete the file from the local storage after successful upload
-        }
+        await acceptObject(dataObjectId, currentPath, newPath)
         registerNewDataObjectId(dataObjectId)
-        addDataObjectIdToCache(dataObjectId, isFileDestinedToLocalVolume)
       }
     } catch (err) {
-      // incluing error from cloud bucket uploads
       logger.error(`Error handling data object ${dataObjectId}: ${err}`)
     }
   }
@@ -258,7 +226,7 @@ export class AcceptPendingObjectsService {
 
     let moved = 0
     for (const dataObjectId of acceptedObjects) {
-      await this.moveToAcceptedLocation(dataObjectId)
+      await this.movePendingDataObjectToUploadsDir(dataObjectId)
       moved++
     }
