--- storage-node/src/services/sync/tasks/index.ts
@@ -1,4 +1,0 @@
-export * from './DownloadTask'
-export * from './StorageProviderSyncTask'
-export * from './DeleteLocalFileTask'
-export * from './ISyncTask'
